Move HiveQueryLanguageStatement to .hql package. Update HiveServerContainer, HiveShellBase, StatementsSplitter and tests.
